code/requirement.txt
flask
langchain
cohere
pinecone-client